Change input file format for easier reading
The previous file format was very messy and confused as it was
riddled with separators.

To make it more user friendly and more alike the physical matrix
card, the cells are separeted by spaces in the same line. This
way it is easier to edit or fix mistakes and to read the file.

Addresses #4